Do not attempt to serialize columns that dont exist.
You cannot serialize attributes that dont exist on a model.

E.g. `Spree::Order.new(foo: :bar)` is exceptional, because foo is not
defined as an attr_writer, or as an attribute on the database table.

Due to upgrading, and changing data models, it's difficult to properly
represent these changes. Instead we should not be serializing objects at
all, but I believe it is done how it is currently for ease of use and
display purposes.